feat(mcp): extract tool input schemas into schemas-lib
Move MCP Zod schemas, JSON-schema conversion, argument parsing, and error
formatting into a dedicated lib module with wrapper re-exports and tests.
